fix html paths on windows
File::Spec returns paths that are native to the operating system and is not meant to build urls in html. This means that on most unixy systems everything goes well, since it generates / separators. However on Windows it generates backslashes, which do not work in HTML. Similar issues will crop up on older mac systems (uses :) or VMS (i don't even know what exactly this does).

File::Spec is simply the wrong tool to use and should be replaced with something that translates any local path to an uri representation, but i don't want to upset the internals too much for now.

As such i've resorted to swapping backslashes with forward slashes on Windows wherever appropiate.